backend/sentinel-server/pyproject.toml
[tool.poetry]
name = "sentinel-server"
version = "0.1.0"
description = "Server for Sentinel, a home security and alert system."
authors = ["Dixon Sean Low Yan Feng", "Keith Neo Kai Si"]

[tool.poetry.dependencies]
sentinel-core = { path = "../sentinel-core", develop = true }
sentinel-opencv = { path = "../plugins/sentinel-opencv", develop = true }
sentinel-desktop-notification-subscriber = { path = "../plugins/sentinel-desktop-notification-subscriber", develop = true }
sentinel-ultralytics = { path = "../plugins/sentinel-ultralytics", develop = true }
sentinel-mediapipe = { path = "../plugins/sentinel-mediapipe", develop = true }
python = "^3.12"
opencv-python = "^4.10.0.84"
aioreactive = "^0.20.0"
expression = "^5.3.0"
fastapi = "^0.115.2"
tortoise-orm = "^0.21.7"
bcrypt = "^4.2.0"
toml = "^0.10.2"
platformdirs = "^4.3.6"
uvicorn = "^0.32.0"

[tool.poetry.scripts]
sentinel-server = "sentinel_server.cli:entry"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
